etcd: Move etcd directory management to run.go
In the etcd package, the only caller of `makeEtcdDir` and
`createEtcdDir` is run.go, so it makes sense for them to live there.
