+++ minion/etcd/network.go
@@ -130,33 +130,6 @@ func runNetwork(conn db.Conn, store Store) {
 	}
 }
 
-func makeEtcdDir(dir string, store Store, ttl time.Duration) {
-	for {
-		err := createEtcdDir(dir, store, ttl)
-		if err == nil {
-			break
-		}
-
-		log.WithError(err).Debug("Failed to create etcd dir")
-		time.Sleep(5 * time.Second)
-	}
-}
-
-func createEtcdDir(dir string, store Store, ttl time.Duration) error {
-	err := store.Mkdir(dir, ttl)
-	if err == nil {
-		return nil
-	}
-
-	// If the directory already exists, no need to create it
-	etcdErr, ok := err.(client.Error)
-	if ok && etcdErr.Code == client.ErrorCodeNodeExist {
-		return store.RefreshDir(dir, ttl)
-	}
-
-	return err
-}
-
 func readEtcd(store Store) (storeData, error) {
 	containers, err := store.Get(containerStore)
 	labels, err2 := store.Get(labelToIPStore)
